separated components into their own files
